requirements.txt
beautifulsoup4==4.12.2
requests==2.31.0
python-dotenv==1.0.0
urllib3==2.0.7
tqdm==4.66.1
openai==0.28.0
pinecone-client==3.0.1
tiktoken==0.5.2
backoff==2.2.1
streamlit==1.31.0
